refactor(annotation): colocate export meta schema and tests with export module
