+++ src/pragmata/core/schemas/annotation_export.py
@@ -1,9 +1,9 @@
-"""Boundary schemas for annotation export records (one per task type)."""
+"""Boundary schemas for annotation export records (one per task type) and run provenance."""
 
 from datetime import datetime
 from typing import Literal
 
-from pydantic import BaseModel, ConfigDict, model_validator
+from pydantic import BaseModel, ConfigDict, NonNegativeInt, model_validator
 
 from pragmata.core.schemas.annotation_task import DiscardReason, Task
 
@@ -107,3 +107,28 @@ class GenerationExportRow(GenerationAnnotation):
 
     constraint_violated: bool
     constraint_details: str = ""
+
+
+class AnnotationExportMeta(BaseModel):
+    """Schema for annotation export run provenance (sidecar to the task CSVs)."""
+
+    model_config = ConfigDict(extra="forbid")
+
+    export_id: str
+    created_at: datetime
+    dataset_id: str | None
+    tasks: list[Task]
+    include_discarded: bool
+    row_counts: dict[Task, NonNegativeInt]
+    n_annotators: dict[Task, NonNegativeInt]
+    constraint_summary: dict[str, NonNegativeInt]
+
+    @model_validator(mode="after")
+    def validate_keys_match_tasks(self) -> "AnnotationExportMeta":
+        """Enforce per-task dicts cover exactly the exported tasks."""
+        expected = set(self.tasks)
+        if set(self.row_counts) != expected:
+            raise ValueError("row_counts keys must match tasks")
+        if set(self.n_annotators) != expected:
+            raise ValueError("n_annotators keys must match tasks")
+        return self
